requirements.txt
python-dateutil==2.8.0
python-dotenv==0.12.0
python-http-client==3.2.7
sendgrid==6.2.1
six==1.12.0
vobject==0.9.6.1